[refactor] cleaning up the files architecture
