+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_i18n/flutter_i18n.dart';
-import 'package:pandemia/navigator.dart';
+import 'package:pandemia/views/home/navigator.dart';
 import 'package:provider/provider.dart';
 import 'data/state/AppModel.dart';
 import 'package:intl/date_symbol_data_local.dart';
